BasketSaverCatalogService/requirements.txt
Django==5.0.3
python-decouple==3.8
mysqlclient==2.1.1
pika==1.3.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1